Storacha-branded validation flow (#417)
**Before Merge:**

* [ ] Set `HOSTED_ZONES` (plural) on `prod`.
* [ ] Merge https://github.com/storacha-network/w3infra/pull/419 (which
this currently includes to get tests to pass).

---

With `NEXT_PUBLIC_W3UP_SERVICE_URL=https://petra.up.storacha.network` in
my local `console`:

![CleanShot 2024-08-19 at 17 33
15](https://github.com/user-attachments/assets/66aa54da-dbf6-47f1-9b1d-a151e9546b9f)

Then…

![CleanShot 2024-08-19 at 17 34
12](https://github.com/user-attachments/assets/03141163-ec6f-447b-bdee-48e3d0ab8abc)

And success!

![CleanShot 2024-08-19 at 17 33
55](https://github.com/user-attachments/assets/fe88b106-daa3-43dd-86ec-a71516bf1c0d)

Or, failure.

![CleanShot 2024-08-19 at 17 34
19](https://github.com/user-attachments/assets/5a8dd394-0f90-429d-a3de-dc6ca9ea3819)

Note that that error message is a dummy message in the test harness, not
something real. Also, I see to have bumped something that made the
Stripe section always appear. Previously, staging/prod was set not to
show it by not providing the right env vars. I'm not sure why that
changed, and whether we want it to.

Closes https://github.com/storacha-network/project-tracking/issues/119

---------

Co-authored-by: hannahhoward <hannah@hannahhoward.net>