fix: consumerTable.get returns customer (#435)
**Context**

While testing the Egress Tracking System in Staging, an error occurred
indicating that the router couldn't resolve the customer:

```
✘ [ERROR] Failed to record egress for space did:key:z6MkkYoZNccuPxDxc8kh6YgPSZwyXCqeS2QYqhtQtYyEKYTt {
    name: 'FieldError',
    message: 'Object contains invalid field "customer":\n' +
      '  - Expected value of type string instead got undefined'
    ...
```

**Issue**

The error arises because the `consumerTable.get` query does not include
the `customer` field in its projection, despite the customer existing in
`consumerTable` and being manually retrievable.

**Fix**

This PR updates the `consumerTable.get` query to ensure the `customer`
field is projected correctly, allowing the router to resolve the
customer as expected. It also fixes the related tests to reflect this
change accurately.