fix(consumerTable): new consumer index (#438)
By creating a new Global Secondary Index (GSI) with a different name, we
avoid any downtime for existing queries that rely on the current
consumer index, and we don’t need to worry about temporarily losing data
in that index.

This PR introduces the `consumerV2` index to the `consumerTable` and
includes the `customer` attribute in the projection. It also updates the
`consumerTable.get` query to use the new index.