fix: remove customer attribute from space diff record (#439)
**Context**
This PR removes the `customer` attribute introduced two days ago in the
space diff record. It doesn't need this attribute, and the table doesn't
support it, which could cause a failure when we attempt to encode/decode
the object. It also adds some additional logs to help with debugging.

There is an error in Sentry that may be related to this change. The
reason behind that is unclear, so this is an attempt to narrow down the
search space and debug.

**Sentry**

![image](https://github.com/user-attachments/assets/b6a8243d-c3b4-405f-a86e-7b9668118716)

**Lambda Logs**

![image](https://github.com/user-attachments/assets/d76dac8a-706d-43ea-a36a-9ea382be4963)