feat: add free trial pricing table to emails when appropriate
We are now offering free trials to users who come referred by other users, so we need to be able to display an alternate pricing table to those users.

This PR adds a check to the "email validated" screen that determines whether to offer users a free trial. It queries the referral service and renders the "free trial" pricing table if the user was referred.

This PR also removes some log lines that I accidentally introduced in a recent PR.