feat: login via OAuth (#463)
refs https://github.com/storacha/RFC/pull/43

This PR adds a GitHub OAuth callback handler. It expects an
`access/authorize` delegation to be passed in the `state` querystring
paremeter. It completes the login flow by issuing a `access/confirm`
invocation on behalf of the user, which in turn issues the requested
delegations and attestations.

If the user is a new user they will be automatically enrolled in the
"trial" plan (`did:web:trial.storacha.network`). This plan does not
require payment details and so their `account` in the `customer` table
is blank.

The billing run will not consider these customers and neither will
egress accounting.

New signups with GitHub accounts must be at least 1 day old.