fix: legacy space claim publishing (#478)
This PR upgrades the `upload-api` dependency, which ensures location
claims for data stored in legacy spaces are sent to the legacy content
claims service (not the indexing service). It also changes the
`space/index/add` hander to publish index claims to the indexing service
_or_ the legacy content claims service, depending on whether the space
is legacy or not. [See the PR for more
details](https://github.com/storacha/upload-service/pull/278).

Most of the change here is reverting the removal of the block index
queue, which feeds into the old E-IPFS indexer. This is because we
decided legacy spaces (spaces provisioned with `did:web:web3.storage`)
go through the existing flows, but when we did, I forgot to consider the
indexing side of things.

Therefore, the changes in the following paths do not need review as it
is not new code:
* `indexer/*`
* `stacks/*`
* `upload-api/external-services/ipni-service.js`
* `upload-api/test/helpers/external-services/ipni-service.js`

depends on https://github.com/storacha/upload-service/pull/278