feat: don't show pricing table to bsky.storage users (#472)
If the `access/authorize` capability is invoked with a fact like
`{appName: 'bsky-backups'}`, don't render a pricing table -instead have
users return to the app to finish the plan setup flow.

We can probably start doing this for any apps, including console, but
let's take it for a spin with bsky.storage.