feat: publish equals claim to appropriate service (#495)
The lambda that runs to publish an equivalency claim now publishes the
claim to the appropriate service.

For legacy spaces the claim goes to the legacy content claims service.
For new spaces it goes to the indexing service.

If we already find an equivalency claim for the content then we don't do
anything!

This is an alternative to
https://github.com/storacha/w3infra/issues/489,
https://github.com/storacha/w3infra/issues/492 and
https://github.com/storacha/w3infra/issues/493 which will give us a
natural ramp up as folks start using the new client library and create
new spaces.