feat: a rough sketch of the Bluesky OAuth implementation
I used Claude to make a rough sketch of the Bluesky OAuth implementation. It has not been thoroughly sanity checked and there are known issues - primarily the in-memory implementations of the session and state stores, but I noticed a few other typechecking issues that may be hiding more bugs.

Pushing this up to support the RFC at [TODO insert link]