chore: change billing monthly run date to the beginning of the month (#506)
To minimize the gap between when the billing period ends and when the
calculations become available.

Currently, the billing process runs on the 28th of each month, covering
the period from the first day of the previous month to the first day of
the current month ( interval is `[from, to)` where `to` is exclusive).
This results in very late invoices for users, since invoices can be
generated on any date they choose. In some cases, the invoice for the
current month may reference usage from two months prior.