fix: provide default success_url to stripe
It looks like Stripe doesn't support the generic "checkout page" they have in their "embeddable pricing table" feature, so `success_url` is required.

The easiest way to resolve this right now is to just provide a default success_url - this makes the `redirectAfterCompletion` option which is honestly good as it was confusing. I'll remove the option from the capability in a future PR.