fix: free trial and "no redirect" fixes (#552)
1) Use coupon codes to implement free trials - free trials and "amount
off" coupons are subtly incompatible with `billing_cycle_anchor` so we
need to do this by giving 100% discounts for "2 months" (actually 1 and
a fractional month because of the `billing_cycle_anchor` - I discussed
this approach with @heyjay44 so we should be good to go forward with it.

2) It looks like Stripe doesn't support the generic "checkout page" they
have in their "embeddable pricing table" feature, so `success_url` is
required.
    
The easiest way to resolve this right now is to just provide a default
success_url - this makes the `redirectAfterCompletion` option which is
honestly good as it was confusing. I'll remove the option from the
capability in a future PR.