fix: proper aggregator service proofs (#545)
These changes require
https://github.com/storacha/upload-service/pull/570 to land first.

Handle `STOREFRONT_PROOF` as a secret rather than an environment
variable.

Also, I noticed that some of the functions that were referencing the
proof were not actually having it passed when defined in the stack, so I
bound the secret to their configs.

Finally, the STOREFRONT_PROOF was renamed to AGGREGATOR_SERVICE_PROOF,
as that aligns with what it actually is. I also removed the storefront
proof from everywhere, based on the assumption that the storefront
service will always be the upload service. Thus, invocations to the
storefront service (the `filecoin/*` family of capabilities) are
self-signed by the upload service and require no proofs.