feat: add a script to update provider field of customer and subscription tables
when we started using the new `did:web:up.storacha.network` provider, new records in these tables ended up with the new provider, but we never updated the old rows. `customer/get` and possibly other capabilities are currently implemented to make it impossible to retrieve subscriptions that don't use the same provider as the service signer, which means some records are simply not available in the new service.

Updating all of these records to use the current provider will, I think, get our system working as expected.

Additionally, we should have an architectural conversation about what "providers" are doing in our system - I don't think they are functioning as intended and would like to either come to a coherent vision of how they should work or get rid of them.