fix: indicate that filecoin proofs are not required explicitly (#577)
Alan was right and this check:
```js
let aggregatorProof
try {
  aggregatorProof = Config.AGGREGATOR_SERVICE_PROOF
} catch {
  // AGGREGATOR_SERVICE_PROOF not bound for this environment
}
```
fails when secrets are fetched from AWS Parameter Store in the `Config`
package's top-level await and the corresponding secret doesn't exist.

However, this only fails if the secret has been bound to the function.
Thus, an alternative to setting `XXX_SERVICE_PROOF = 'none'` is to bind
the secret conditionally.

This PR does three things:
1. Introduces a new `FILECOIN_PROOFS_NOT_REQUIRED` variable which will
prevent Filecoin pipeline-related proofs (the aggregator service and
deal tracker for now) to be bound to relevant functions. This makes
passing proofs the default, and we'll need to set this var for hot
network envs.
2. Binds the aggregator service proof to the ucan invocation router
function, which I forgot to do in my past PR.
3. Uses a try-catch block to check whether `DEAL_TRACKER_SERVICE_PROOF`
is set, to make it consistent to how this is done for
`AGGREGATOR_SERVICE_PROOF`. This allows removing the
`DEAL_TRACKER_SERVICE_PROOF = 'none'` from hot network environments.